Handle pressing cancel in file dialog
If cancel was pressed, nothing would be written to the buffer.
We now check for this by having the first byte of the buffer being set
to zero by default, so that if no path is selected, there will still be
a valid null terminator, within which there will be a valid UTF-8
string.