mshtml, edge: speed up `Content::Html`
Previously `sscanf` is used in decoding the url-encoded html document.
But `sscanf` is quite expensive. Loading a larger document (~800KiB) has
a noticeable delay.

And mshtml is giving the third argument to `sscanf` a `char*` when it
should be an `int*`. It is very problematic.

This commit replaces them with our own hex decoding function.